Update patch set 2
Patch Set 2: Code-Review-1

I think it's better not to do this on stable-2.12.  I've uploaded a change for master that includes a schema migration:

https://gerrit-review.googlesource.com/82352

If this might be critical for users of 2.12.x we can add a "Known issues" section to the release notes with instructions to manually increase the column size.

Patch-set: 2
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1