Update patch set 1
Patch Set 1:

> After reading the bug I understand why this solves the problem. What I don't yet understand is why this is better than making all pages that serve the PolyGerrit loader HTML thingie just never be cacheable (which IIRC Andy was already doing?).

I'm not sure this is a better solution, tbh - it just seems logical to serve cookie if the user is signed in.

If auth expires before the page is ever refreshed, this would require loading home page in child window with sole purpose of getting the cookie, which is a bit more than minimally required.

Not feeling strongly though.

Patch-set: 1