Update patch set 3
Patch Set 3:

(7 comments)

No need to overthink the tests. Most of the tests you marked are a basic sanity check of "if we set these fields the output looks like this," which IMO is useful when you're defining a new serialization format. JSON is not novel so many of them don't need to be replicated.

Patch-set: 3
Label: Verified=0