Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Hitting login brings me to a 404 page. Perhaps due to the changes in gr-router.

I am unable to reproduce a regression between my branch and master. This is what I did to compare the results:

On master:
1. Run `./polygerrit-ui/run-server.sh`
2. Open localhost:8081
3. Hit Sign in
4. See a white page with in the top left corner "404 page not found"

On my branch:
1. Run `./polygerrit-ui/run-server.sh`
2. Open localhost:8081
3. Hit Sign in
4. See the default Gerrit layout, but in the middle of the screen:
¯\_(ツ)_/¯
404
Back to home

E.g. I do not observe a regression, since Sign in does not appear to work on master either (for me). However, the message is a bit friendlier now with the white page gone and a nicer 404 page.

Is there a step missing to make Sign in work on master?

Patch-set: 4
Label: Verified=0