Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > (2 comments)
>  > 
>  > How can you search for changes with no assignee? That seems like an
>  > important use case.
> 
> Agree. Plan to add a "is:assigned" operator in a follow up change.

The problem is if you are storing an integer field in the index, there is no way to write a query that says "has no value for this field". So you have to change your field format for this field if you want to support this type of query.

Patch-set: 4