Update patch set 1
Patch Set 1: Code-Review-1

Using a set for the return type from ChangeNotes makes sense, because they're account IDs, and that allows us to dedupe. However I think it's silly to return as a set from this REST API handler:

* AccountInfo does not implement hashCode and equals
* JSON doesn't have sets, so Gson converts it to a list anyway

If you wanted to implement hashCode and equals on AccountInfo, maybe this would be ok, but it really just doesn't seem to me like having this return a set is that useful.

Maybe if some user of the extension API really cares about it then we can return a Map from the extension API (not the REST API). But I doubt it matters.

Patch-set: 1
Label: Code-Review=-1