Update patch set 2
Patch Set 2:

> The original change in this case is the change being merged in.
 > 
 > Say we have change A, and change B which is the merge of change A.
 > If we update the patchset on change A to A', change B is still the
 > merge of change A.
 > If we actually want change B to be the merge of A', we'll need to
 > update change B to change B'.

This seems like a pretty low-level way of accomplishing that goal, because the caller has to go look up what A' is so they can pass it in the MergeInput. A higher-level way to accomplish this goal would be to make rebase DWIM.

That said this is probably better than nothing. I will just have to think about how to name this so that it's understandable. I don't think "update change by merge" is doing it for me.

Patch-set: 2