Update patch set 1
Patch Set 1:

(1 comment)

I'm -0 on this kind of change. Specifically, StreamSupport.spliterator is so ugly that it's not a readability win over FluentIterable.

Guava 21 or so will have a Streams utility class so that the ugly StreamSupport.spliterator call can be replaced with a nicer Streams.stream(Iterable). Of course there is no real harm in doing this twice (FluentIterable -> StreamSupport -> Streams), but it's churn.

Taking a step back, IMO GroupCache should just return List.

Patch-set: 1