Update patch set 1
Patch Set 1:

> > WDYT? Any better approaches?
 > 
 > The only alternative I can think of is to construct the Repository
 > and RevWalk inside the listener, using the given project name and
 > commit.

I thought of that and wondered if the newly received objects would be in the project at all at the point this check runs. I think in case of push command, the newly received objects are still kept in separate "pool", but I haven't tried verifying this yet.

Patch-set: 1