Update patch set 12
Patch Set 12: -Code-Review

also, how can it be that there is a production case that is broken, but we have no CI failure? 

Don't we have a test that runs the actual production WAR and then runs curl to check that it is basically working?

Patch-set: 12
Reviewer: Gerrit User 1024147 <1024147@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=0