Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Have you seen anything like this on other dialogs? I'm no mobile expert, but I was addressing a similar issue for the cherrypick dialog. https://gerrit-review.googlesource.com/c/90372/

This issue is specific to the way that android handles the window with the keyboard up. 

So far, in vertical orientation mode, I did not see other overlays with this issue (I looked at many others as well). We could in theory add the scrollable class to gr-overlay itself, however, the size of the content inside it (to make it appear full width/height) is not handled by the overlay but rather what's inside it (in this case gr-reply-dialog).

In horizontal orientation mode, everything is kind of screwed up for many textareas, and we would have to work with it on a case by case basis. I chose not to worry about that for this change since in my opinion, if they can use it in vertical mode, at least people can get their work done. 

I think it would be worth doing a deeper dive into this later, though there is a larger iOS issue that I submitted a bug for relating to the overlay, and without resolving that first, I don't know if investing more time is worth it. I think that should be the priority and some other things may follow.

Patch-set: 1