Update patch set 4
Patch Set 4:

> >  > it _is_ in fact missing a dependency, and
 > >  > you shouldn't be allowed to merge it?
 > Why not? Is there more valuable feedback to be be given than
 > "depends on stale patchset of merged issue, consider rebasing?"
 > Assuming not, consider this wrt to busy repositories with 100+
 > commits per day (like Chromium). There, effectively every change
 > has to be rebased before landing (under Rebase-* strategy), and
 > asking developers to click "rebase" and retry "submit" doesn't seem
 > necessary to me.

The ps b,1 is depending on ps a,1 and change a has new patchsets.
* The verified flag is no longer valid for b since it's applied to b,1 and in order to submit you need to submit b,2 which is not guaranteed to have the same content as b,1 simply because it's not the same commit.
If the rebase-submit is done in the same transaction you can be sure that base -> b,1 have the same delta as new-base -> b,2. But once you do them one by one you can't tell.
What if we have to changes a,1 and b,1 it builds and you get a Verified+1 on both changes but then you change direction of a completely and b,1 no longer builds if you apply it on top of a,2.
Then you submit a,2.
Should you really under these circumstances be able to submit b,1? Would you even want it to be submitable?
And how do you separate this from an instance where you just decided to submit them one by one?

The dependency of b,1 is no master -> b,1 it's a,1->b,1.

Patch-set: 4
Label: Verified=0