Update patch set 8
Patch Set 8:

(7 comments)

I've a few nits and such.

In the future, would you consider making this a series of related changes instead of one big one? It's a bit daunting to do an 800-line review. I know it's ultimately the same amount of work for the reviewer anyway, but several 100-line changes are psychologically much more appealing :)

Patch-set: 8
Label: Verified=0