Update patch set 5
Patch Set 5: Code-Review+1

With this most recent patch set, the API portion of resolvable comments is finished, EXCEPT for one edge case.

In PolyGerrit, comments are no longer 'threaded', insofar as each thread created as an inline comment is sorted entirely by chronology. This means that the state of resolution for the thread is equivalent to the state of the boolean field on the most recent comment. 

In GWTUI, comments are still threaded. Resolvable comments are not supported, and comments with no unresolved field supplied have that field populated with the value of the 'in_reply_to' comment.

Sometimes, this in_reply_to comment != the last comment. Therefore, if a user on GWTUI replies to a comment where unresolved = false, but lower down in the thread someone on PG set unresolved = true, the state will inadvertently be modified to have unresolved = false again.

This has proven to be a very hard issue to solve, mainly due to the black box that is the GWTUI. Removing threading from GWTUI as well would be a substantial task on a legacy UI.

There are two possible solutions to this issue:
- Modify the behavior of comments with no unresolved field supplied, i.e. to have them default to unresolved = true
- Deal with this edge case, which *likely* won't happen very often, as users are migrating to PG as whole teams... In other words, it is very unlikely that two users would be communicating on the same change with different UIs.

Thoughts?

Patch-set: 5
Label: Code-Review=+1