Update patch set 3
Patch Set 3:

(3 comments)

The code LGTM, I have one question on the behavior.

Is it intended that if I open the [Reply] dialog, remove a reviewer and then [Cancel] the dialog, the reviewer is still removed? I'd have expected that there would be a pending-remove state to make it symmetric with adding.

Patch-set: 3
Label: Verified=0