Update patch set 4
Patch Set 4: PolyGerrit-Review+1

(2 comments)

> Is it intended....

It's something that I've noticed, acknowledged, and ultimately decided to ignore, mostly because there's no way (that I can see) on the backend for batch-removing reviewers. Instead, this functionality provides symmetry with the removal of fields in change-metadata.

It's my opinion that we should do away completely with the Owner/Reviewers/CC fields in the reply dialog and rely entirely on inline autocomplete addition of reviewers in change-metadata... But that's a discussion for another time, especially when we don't have pending deadlines.

Patch-set: 4