Update patch set 4
Patch Set 4:

Thank you for your feedback so far. I've appreciated it very much. Is there anything I can do so that this change can be merged soon?

When I looked at all of the comments again, my impression was that only our different opinions regarding the class TreeCreator remain. Before I start to mess around with this class, I would like to find a compromise which is acceptable for all of us.
As I explained, the suggestion "new TreeModifier(baseTree).apply(repository, modification);" doesn't feel good to me and I'd rather not provide code which I disapprove of. If you insist that the specification of a TreeModification must be enforced (beyond a simple check) and that it should be only one (instead of a list), I can offer to transform the class into a utility class having just one public static method. The result would be similar to "TreeModifier.createNewTree(repository, baseTree, modification)". It's a solution I wouldn't be entirely happy with but I could live with it. Would this variant be acceptable for you?

Patch-set: 4
Label: Verified=0