Update patch set 4
Patch Set 4:

> Thank you for your feedback so far. I've appreciated it very much.
 > Is there anything I can do so that this change can be merged soon?

Thank you for detailed comments. I haven't reacted on them, because I agree with you. Just let it all as is. I'm happy to move forward with this series, if you add a comment/TODO on List<TreeModification> treeModifications declaration in TreeCreator.java, that only one item is currently used, and it would be extended in near future, as suggested in my last comment.

Patch-set: 4