Update patch set 1
Patch Set 1:

The main idea behind this, is when we load comments we will add __commentSide as an attribute to all comments, so if we end up merging right and left in a unified view, we can still create separate threads for both of them.

The more complicated part was making sure that when comment/drafts are added, and saved, that they also get the correct attribute, so if you switch views from unified/side-by-side/etc, things still render properly.

A few outstanding questions:
- Do we still need 'side' (Parent/Revision) ? I kept all that in there, but it might be overkill. I think some of it is needed though, not exactly sure.

Patch-set: 1
Reviewer: Gerrit User 1030207 <1030207@173816e5-2b9a-37c3-8a2e-48639d4f1153>