Update patch set 1
Patch Set 1:

> I would place the UTC offset in the tooltip text if developers are requiring that specifically. For instance, in Rietveld it’s displayed as 2017-01-31 18:33:43 UTC, which is fine in a tooltip.

Do you mean to include the UTC offset in the tooltip (with the seconds) and leave the display text as is?

UTC offsets are not as pretty as timezone codes, but IMO they strike a good balance between conveying the relevant info, and not needing to load bulky timezone data that needs to be updated periodically.

Patch-set: 1
Reviewer: Gerrit User 1026519 <1026519@173816e5-2b9a-37c3-8a2e-48639d4f1153>