Update patch set 4
Patch Set 4:

Hi, would it be possible to allow the admin to allow all users to delete there own changes if they wont. (not allowing other users to delete other users patches except from admins for that but allow normal users delete there own patches if the admin has configured it to allow this please?

Patch-set: 4