Update patch set 3
Patch Set 3: PolyGerrit-Review-1

(1 comment)

> Patch Set 3: PolyGerrit-Review+1
> 
> (1 comment)
> 
> I too feel weird about the extra vertical.
> 
> > Say, right before the date?
> 
> What about below the date in the "expanded" layout of the last comment in the thread?

I'm not sure that that's right either, because it'll create an extra vertical row in the comment (or force the comment content to reflow).

I think that the best place to put this is where it is, but to only show it when the buttons themselves are also shown. The visual indicator of the different color for the thread should suffice in most cases, and the state of the thread itself really doesn't matter in any case except one in which you're able to reply. How does that sound?

Patch-set: 3
Reviewer: Gerrit User 1029297 <1029297@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: PolyGerrit-Review=-1