Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 1: PolyGerrit-Review+1
> 
> (1 comment)
> 
> This is really nice.
> 
> I predict that we will soon wish for copy-to-clipboard buttons elsewhere in the app (e.g. HTTP password generation). Ultimately, we may wish for a clipboard component with an interface similar to CBN-COPY-CLIPBOARD [1].
> 
> That doesn't necessarily need to block this change, though. What do you think?
> 
> * [1] https://github.com/CBN-IT/cbn-copy-clipboard

Yeah that's an interesting idea. I think we could create an element pretty easily. I don't love the way they do the behavior, if we already have an element on the page with the contents we need (you pass in text, it creates a new element adds the text, and then uses that to copy to clipboard), but it does make things seem simpler. I think it could be worthwhile to determine whether or not we want to be able to pass a string to copy or an html element.

Patch-set: 2
Reviewer: Gerrit User 1030207 <1030207@173816e5-2b9a-37c3-8a2e-48639d4f1153>