Update patch set 27
Patch Set 27:

> Hi, will this also migrate external ids for non notedb installs,
 > like mysql installs please?

Not in this change, but in the next one. For there reasons, it's done in two steps. In next change, all external ids are imported from ReviewDb to NoteDb and the table in ReviewDb is deleted.

Patch-set: 27
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>