Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This can cause a significant flicker. Is there a way to check for the index first and not hide so many messages in the initial render?
> 
> Maybe that doesn't need to block this change, though.

As far as I can tell, there doesn't seem to be an easy way to do this. This function is (most often) called by the function _maybeScrollToMessage in gr-change-view, which itself has dependencies on the message-list loading. I could look into seeing if those dependencies can be unrolled somehow if you think it's worth the time, but it's not unusual for UIs to flicker in cases like these (when scrolling to active content has dependencies) and due to the way we use this.splice to add to the comments array, you aren't reloading anything.

Patch-set: 1
Reviewer: Gerrit User 1029297 <1029297@173816e5-2b9a-37c3-8a2e-48639d4f1153>