Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > comment-side is either left or right, and side is either parent or revision.  > For this, we need to know if it is for the parent or not.
> 
> This is super-confusing, can we somehow make this more explicit?

I agree it is confusing. I think commentSide is more of what I would expect 'side' to be, and 'side' I would expect to be a boolean isParentRevision or something to that effect, but the API calls it 'side' with the string 'PARENT' or 'REVISION'.

See previous comments here:
https://gerrit-review.googlesource.com/c/95811/1/polygerrit-ui/app/elements/diff/gr-diff-builder/gr-diff-builder.html#213

Patch-set: 1
Reviewer: Gerrit User 1030207 <1030207@173816e5-2b9a-37c3-8a2e-48639d4f1153>