Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Do we need specific endpoints for this?  Instead can we use the existing add reviewer endpoint, and make it so that when a reviewer (or CC) is added with only email address, it falls back to adding as an unregistered CC (if allowed) when there is no account associated to that address?

Patch-set: 1
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1