Update patch set 1
Patch Set 1:

Can this wait until after 4.7.0 is released?  I plan to update to 4.7.0 on stable-2.14.

If we do this upgrade here it'll just be confusing to have master with a version that looks older (but is actually a later snapshot off JGit's master).

Patch-set: 1
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>