Update patch set 1
Patch Set 1:

> Why not making everything consistent with:
 > 
 > showAssignee = true/false (global enabling the visibility)
 > showAssigneeInChangeScreen = true/false
 > showAssigneeInChangesTable = true/false
 > showAssigneeInSearchOracle = true/false
 > 
 > By default, everything is true except the changes table.

IMO it doesn't really make sense to disable just any of those, but you felt that it was annoying to have the column in the change table, and that, to me, would be the only thing that you would have any reason to disable without disabling the whole feature.
Disabling just one of change screen or search oracle makes so little sense to me that it's IMO not worth the verbosity in the Documentation, let alone the cluttering of the source.

Patch-set: 1