Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> > Why not making everything consistent with:
>  > 
>  > showAssignee = true/false (global enabling the visibility)
>  > showAssigneeInChangeScreen = true/false
>  > showAssigneeInChangesTable = true/false
>  > showAssigneeInSearchOracle = true/false
>  > 
>  > By default, everything is true except the changes table.
> 
> IMO it doesn't really make sense to disable just any of those,

Of course not, as you guys have developed this feature for your use-case, you wouldn't like to disable it afterwards :-)

My feedback was coming from the observation that neither in Gerrit nor in GerritHub the feature is highly used: the % of changes with an assignee associated is very little.

> but you felt that it was annoying to have the column in the change table, and that, to me, would be the only thing that you would have any reason to disable without disabling the whole feature.

Again, the feature *cannot be* disabled as you agreed. It is just a matter of where to display the assignee or not. But the feature is there and cannot be turned off.

> Disabling just one of change screen or search oracle makes so little sense to me that it's IMO not worth the verbosity in the Documentation, let alone the cluttering of the source.

Fair enough, why not then just introduce the "showAssigneeInChangesTable" with "false" as default value and leave the feature always enabled? Giving the impression to be able to disable it is misleading.

Patch-set: 2
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>