Update patch set 5
Patch Set 5:

> I haven't downloaded this to try it out yet. Does it just tell the user to reload? Would be good to have a link to reload for them.

Oops. Missed this message in my earlier pass.

I've added a reload link to the reply dialog.

To add such a link to the toast, we will need to generalize the "action" aspect of auth toasts to other alert types. I'd like to do this part in a different change.

Patch-set: 5