Update patch set 2
Patch Set 2:

For debian there may be problems with nodejs 6.x as i understand that there is already a node command on debian which nodejs 6 renames the command too. On a wikimedia instance i found this problem with jessie. We may want to create a bashrc file which translates the command to nodejs? or create a symblink from a nodejs command to node command.

Patch-set: 2
Reviewer: Gerrit User 1029953 <1029953@173816e5-2b9a-37c3-8a2e-48639d4f1153>