Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > I like it! What do you think about adding some hover text in the "warn" state to help explain what is happening?
> 
> That'd be a reasonable feature, I'm not sure what it should say, though, since "committed"-ness isn't really a user-exposed term. My thinking was just to make a subtle hint.
> 
> > it feels like a lot of extra functionality has already been bolted on to the autocomplete.
> 
> What extra functionality do you mean? It doesn't seem to me like much is there. 
> 
> This feature also feels to me like the logic is all internal to the autocomplete. What's the benefit of splitting it between the two elements?

We've already added a ton of properties to the gr-autocomplete, most of which are only used in one use case. This feels like a code smell to me.

If there are lots of unique use cases for behavior that's desired on, say, blur, or commit, or whatever, perhaps we ought to be listening for the event in the parent element and keeping this element simple.

Patch-set: 2
Reviewer: Gerrit User 1029297 <1029297@173816e5-2b9a-37c3-8a2e-48639d4f1153>