Update patch set 3
Patch Set 3:

> Also, why is this being added? Is there a bug?

I didn't want to link this directly to the bug, but this is partly in response to issue 6350.

We've also seen a number of issues where users are confused about whether the text that was entered into the autocomplete (but not committed) ends up getting sent in the reply. This is just intended as a subtle hint that the text is only staged in the autocomplete.

> When would we not want a warning for uncommitted text in the account-entry?

I'm confused by this question. With this change, the account entry always warns after being blurred.

Patch-set: 3
Reviewer: Gerrit User 1026519 <1026519@173816e5-2b9a-37c3-8a2e-48639d4f1153>