Update patch set 2
Patch Set 2:

(3 comments)

Thanks for getting the project index started. This change looks already pretty good.

 > One more code review question:
 > What are the tests I need to add for this change? Do not see any
 > test added for the group index schema change(https://gerrit-review.googlesource.com/c/93484/).

It's a whole topic that added support for the group index [1]. Topic means that it's a series of changes that are submitted together. The tests for the group index come later in the series [2]. The tests are based on the REST API and the query support which needs to be added first.

[1] https://gerrit-review.googlesource.com/#/q/topic:group-index+(status:open+OR+status:merged)
[2] https://gerrit-review.googlesource.com/93500

Patch-set: 2
CC: Gerrit User 1003935 <1003935@173816e5-2b9a-37c3-8a2e-48639d4f1153>