Update patch set 1
Patch Set 1:

(1 comment)

> Can you elaborate why are we adding new test features on very old
 > stable branch that is using Java7?

Hugo is changing the high-availability plugin to have its configuration in its own file (i.e. etc/high-availability.config rather than etc/gerrit.config) but it has a test (CacheEvictionIT) that is setting config values using @GerritConfigs.  This won't work if the config is moved to the plugin's own file.

I'm adding this on stable-2.13 since high-availability is targetted at 2.13.x

An alternative would be not to do that config change on the plugin for 2.13.x but punt it to its stable-2.14 branch.  Then we can also move this to core gerrit stable-2.14.  The down side of doing that is, if anyone uses it on 2.13 they then need to migrate the settings on upgrade to 2.14.

Also open to other suggestions of how to fix the problem with CacheEvictionIT.

Patch-set: 1
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>