Update patch set 1
Patch Set 1:

(1 comment)

The best design for a live region is to have a permanent element that's a live region - like a container of the element that's hidden or shown. Then hide or show elements within that container, and they'll be spoken automatically when they appear.

Because of the way they're handled by assistive technology, creating a live region and having it speak automatically is kind of a race condition, it's not surprising that it won't work consistently.

Patch-set: 1
CC: Gerrit User 1046187 <1046187@173816e5-2b9a-37c3-8a2e-48639d4f1153>