Update patch set 3
Patch Set 3:

This is at odds with the ongoing PermissionBackend refactoring. I suppose I don't necessarily object to this in the short term, but be aware that all the RefControl#can* methods are going away in the near future. (Actually, given that it's going away soon, I'm probably more likely to approve this in the short term without bikeshedding the public API :)

In the long term, we need to figure out a way to make this kind of information available in the PermissionBackend interface. Given the current model of "void check(FooPermission) throws AuthException", the only way to wedge it in would be as the AuthException message.

Patch-set: 3
CC: Gerrit User 1010008 <1010008@173816e5-2b9a-37c3-8a2e-48639d4f1153>