Update patch set 2
Patch Set 2:

> This change breaks compilation of the gerrit-rest-java-client. All my 
> implementations extend the corresponding 'NotImplemented' classes. 

Urs, it seems to me that we are breaking the gerrit-rest-java-client at least occasionally. Could you in future make an effort to test the gerrit-rest-java-client against new release candidates so that we notice such breakages earlier? Giving people an opportunity to test an upcoming release is actually the reason why we create release candidates.

Maybe even better if you could cooperate with Luca to add a build job for the gerrit-rest-java-client to our CI server so that breakages can be noticed automatically.

Patch-set: 2