Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > This change breaks compilation of the gerrit-rest-java-client. All my 
> > implementations extend the corresponding 'NotImplemented' classes. 
> 
> Urs, it seems to me that we are breaking the gerrit-rest-java-client at least occasionally. Could you in future make an effort to test the gerrit-rest-java-client against new release candidates so that we notice such breakages earlier? Giving people an opportunity to test an upcoming release is actually the reason why we create release candidates.
> 
> Maybe even better if you could cooperate with Luca to add a build job for the gerrit-rest-java-client to our CI server so that breakages can be noticed automatically.

Edwin, thanks for fixing it so fast.

This change did not break anything for any user of the gerrit-rest-java-client. It did just break my manual update procedure which I do occisonally.

Manual update procedure means: copy most of the classes from module 'gerrit-extensions-api' into the gerrit-rest-java-client project and apply some patches.

I know that this is not the greatest way to do this. Reasons for this way are:
* gerrit-rest-java-client still provides Java 6 support while Gerrit depends on Java 8 now.
* The module 'gerrit-extensions-api' is not available als Maven dependency AFAIK.
* There are some minor additions to the API which are not ready (yet) to get merged into Gerrit.

I do not have an idea how we could automatically detect such breakages at the moment...

Patch-set: 2