Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Is the gerrit.serverId setting a permanent feature or is it likely to get removed/replaced during migration to note db?
> 
> If it's permanent it should probably be mentioned in the documentation.

It's permanent, at least until we come up with a better scheme as described in the commit message of this change (which we haven't made progress on). I still have several undocumented config options to take care of before NoteDb becomes officially supported, but thank you for reminding me that this is on the list.

Patch-set: 5