Update patch set 1
Patch Set 1:

(1 comment)

There is a risk to caching these forever: if the large object threshold is tunable (depends where the stack trace is coming from, some are some aren't IIRC), then restarting the server with a larger threshold would not result in clearing the cache of tombstones for failures that would succeed under the new threshold. This might be confusing to users/admins, if they expected increasing the threshold to have a visible effect on changes that didn't work before.

On possibility is to change the cache key to include the threshold. Not sure that is worth it though, and you can always do it later. In the short term admins could address user confusion by manually removing cache entries.

Patch-set: 1
Reviewer: Gerrit User 1010008 <1010008@173816e5-2b9a-37c3-8a2e-48639d4f1153>