Update patch set 2
Patch Set 2: Code-Review+1

We are using this in production, and it helps a lot. Even though a new release wont happen, it may be a good idea to merge this backport to make it easier for people to find and build it when needed.

Patch-set: 2
Reviewer: Gerrit User 1021592 <1021592@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=+1