Update patch set 1
Patch Set 1:

I hesitate on this change because (unlike in the GWT UI) the rebase and cherry-pick buttons are supposed to be independent of the selected patch -- they normally only work on the latest one to avoid confusion.

It could be doubly confusing if they depend on the selected patch, but only if the patch is an edit. That's a bit subtle.

Maybe we can just add a tooltip to these buttons before you click that, although you have selected some patch other than the latest published patch (edit or not), the action only applies to the latest one.

What do you think?

Patch-set: 1
Reviewer: Gerrit User 1026519 <1026519@173816e5-2b9a-37c3-8a2e-48639d4f1153>