Update patch set 1
Patch Set 1:

> > I hesitate on this change because [..]
>
> [..] TLDR, I suppose, is that these actions don't exist in (what I consider to be) the CUJ of an in-browser edit, and so I'd like to hide them. [..]

For some context, keeping the revision actions locked to the latest patch set was a very deliberate decision [1] -- and broadly speaking, I think it was a good choice. It makes little sense to rebase an older patch set, and the button to cherry-pick an older patch is kind of weird, and hasn't really been missed, so far as I've seen.

However, that decision was made before PG would show edits, so I can understand if it's become outdated. I'm just trying to think of a way that would be consistent with the existing pattern and doesn't surprise users with buttons that disappear for subtle reasons. (I have some other ideas in this direction.)

[1] https://gerrit-review.googlesource.com/c/gerrit/+/84091

Patch-set: 1
Reviewer: Gerrit User 1026519 <1026519@173816e5-2b9a-37c3-8a2e-48639d4f1153>