Update patch set 26
Patch Set 26:

(1 comment)

@Becky Siegel i found an issue on chrome where you refresh the page, it would hide everything, even if your a memeber, the issue does not happen on safari. So fixing that will be out side of this change as it may be a bug in the actual chrome browser.

Also another issue found is that the delete button seems to always be shown even if your not a member. This also seems like a bug as i am hiding it, but it still wont hide. I presume this is because i am using dom-repeat. Any ideas how to fix this? Fixing that can be in a seperate change, if not easy to fix, since this fixes allowing admin's to edit groups and also fixes it for members to. Includes some fixes :).

Patch-set: 26